Improve screen reader usability of API tour
When using a screen reader, the flow of the tour is lost when the result
is generated. As a non-visual user, it is impossible to know that the
results section has been added.

Also fixes an issue whereby the "Run this JavaScript example" button
does not correctly display as disabled after it has been clicked.