Fix incorrect example responses
The example responses being displayed in the API endpoint docs was
actually a JSON serialised view of the entire spec model for the
response schema rather than an example response.

This fix will generate an example for all responses where a schema is
specified. It will determine the value for each property using the
following precedence (if attribute is not set it will go to the next):
1. `example` value
2. `default` value
3. first `enum` value
4. sensible defaults based on the `type` and `format` where specified
5. `type` value
Using this method all properties can be represented in the example so
long as the API spec is valid.